Add script for extracting load time series in units kW/h
Add script for extracting load time series in units kWh/h for existing load points in the CINELDI reference grid.
In addition the handling of normalized load time series (in load_profiles.py) had to be generalized to be able to extract time series for an arbitrary set of days of the year.