Add script for creating grid with a load snapshot
Add the script create_grid_with_load_snapshot.py for creating a version of the grid data set for a certain operating state,  obtained for a "snapshot" for a given day and hour of the year from the load demand time series. In this version, the new grid data is just being outputted as a new bus .csv file.